Move Login and Routine Editor to pages. Move ExerciseList, ExerciseItem, and CreateExerciseModal to ExercisePanel
