+++ client/src/components/ExercisePanel/ExerciseList/ExerciseList.js
@@ -1,5 +1,5 @@
 import React, { PropTypes } from 'react'
-import { ExerciseItem } from 'components'
+import ExerciseItem from '../ExerciseItem/ExerciseItem'
 import './styles.css'
 
 const propTypes = {
