+++ client/src/pages/index.js
@@ -1,1 +1,3 @@
+export { default as Login } from './Login/Login'
+export { default as RoutineEditor } from './RoutineEditor/RoutineEditor'
 export { default as RoutinesPage } from './RoutinesPage/RoutinesPage'
